test: moviendo tests de integracion a app/
