+++ app/src/test/java/poliplanner/test/integration/ExcelServiceIntegrationTest.java
@@ -1,0 +1,92 @@
+package poliplanner.test.integration;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assumptions.assumeTrue;
+
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.time.LocalDate;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.Rollback;
+
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import poliplanner.excel.ExcelService;
+import poliplanner.excel.sources.ExcelDownloadSource;
+import poliplanner.repositories.CareerRepository;
+import poliplanner.repositories.SheetVersionRepository;
+import poliplanner.repositories.SubjectRepository;
+
+@SpringBootTest
+@Transactional
+@Rollback
+@RequiredArgsConstructor
+class ExcelServiceIntegrationTest {
+    final private ExcelService excelService;
+    final private SheetVersionRepository sheetVersionRepository;
+    final private CareerRepository careerRepository;
+    final private SubjectRepository subjectRepository;
+
+    private static final String EXCEL_DOWNLOAD_URL = "https://www.pol.una.py/wp-content/uploads/Horario-de-clases-y-examenes-Segundo-Academico-2024-version-web-19122024.xlsx";
+
+    @Test
+    void testExcelDownload() throws Exception {
+        // Tratar de descargar un excel conocido y ver si es que funciona
+        // correctamente
+        try {
+            ExcelDownloadSource source = new ExcelDownloadSource(EXCEL_DOWNLOAD_URL, "test_file", LocalDate.now());
+            Path file = source.downloadThisSource();
+
+            assertNotNull(file);
+            assertTrue(Files.exists(file));
+            assertTrue(Files.size(file) > 0);
+        } catch (Exception e) {
+            System.out.println("Test ignorado. No se puede conectar con el url de descarga");
+            assumeTrue(false);
+        }
+    }
+
+    // @Test
+    // @Tag("integration")
+    /**
+     * Test de integración que verifica el proceso completo de parseo de un archivo
+     * Excel y la
+     * persistencia de sus datos en la base de datos.
+     *
+     * <p>
+     * Flujo verificado: 1. Lectura y parseo del archivo Excel. 2. Persistencia de
+     * la versión del
+     * archivo (SheetVersion) con la URL asociada. 3. Creacion y persistencia de
+     * entidades Career y
+     * Subject a partir del contenido del Excel.
+     *
+     * <p>
+     * Este test asume que el archivo testeado contiene datos validos y
+     * estructurados
+     * correctamente.
+     */
+    /*
+     * void testExcelParsingPersistency() throws Exception {
+     * Path testExcel = Paths.get("src/test/resources/testExcel.xlsx");
+     * String dummyUrl = "URL";
+     * 
+     * excelService.persistSubjectsFromExcel(testExcel, dummyUrl);
+     * 
+     * // Verificamos que se haya creado la version
+     * SheetVersion version =
+     * sheetVersionRepository.findFirstByOrderByParsedAtDesc();
+     * assertEquals(dummyUrl, version.getUrl());
+     * 
+     * // Verificamos que se hayan creado carreras
+     * List<Career> careers = careerRepository.findAll();
+     * assert (!careers.isEmpty());
+     * 
+     * // Verificamos que se hayan creado materias
+     * List<Subject> subjects = subjectRepository.findAll();
+     * assert (!subjects.isEmpty());
+     * }
+     */
+}
